Fold precision coverage into existing test suites
