+++ packages/platform/bun/test/BunHttpPlatform.test.ts
@@ -1,3 +1,4 @@
+import type * as BunFileSystem from "@effect/platform-bun/BunFileSystem"
 import * as BunHttpPlatform from "@effect/platform-bun/BunHttpPlatform"
 import { assert, describe, it } from "@effect/vitest"
 import * as ByteSize from "effect/ByteSize"
@@ -5,6 +6,8 @@ import * as Effect from "effect/Effect"
 import type * as HttpBody from "effect/unstable/http/HttpBody"
 import * as HttpPlatform from "effect/unstable/http/HttpPlatform"
 import * as HttpServerResponse from "effect/unstable/http/HttpServerResponse"
+import { afterEach, beforeEach, vi } from "vitest"
+import { testFileResponsePrecision } from "../../node-shared/test/HttpPlatform.test-utils.ts"
 
 const readBody = (body: HttpBody.HttpBody) => {
   assert.strictEqual(body._tag, "Raw")
@@ -32,7 +35,7 @@ const readResponse = (response: HttpServerResponse.HttpServerResponse) =>
       })
   )
 
-describe("BunHttpPlatform", () => {
+describe("BunHttpPlatform", { concurrent: false }, () => {
   for (
     const { name, offset, bytesToRead, expected } of [
       { name: "clamps bytesToRead beyond EOF", offset: 1, bytesToRead: 10, expected: "bcd" },
@@ -108,5 +111,47 @@ describe("BunHttpPlatform", () => {
         { exact: yield* readBody(exact.body), empty: yield* readBody(empty.body) },
         { exact: "import", empty: "" }
       )
+    }).pipe(Effect.provide(BunHttpPlatform.layer)))
+})
+
+const precision = vi.hoisted(() => ({ enabled: false }))
+
+vi.mock("@effect/platform-bun/BunFileSystem", async (importOriginal) => {
+  const original = await importOriginal<typeof BunFileSystem>()
+  const { Layer, Effect } = await import("effect")
+  const { fileSystemLayer } = await import("../../node-shared/test/HttpPlatform.test-utils.ts")
+  return { layer: Layer.unwrap(Effect.sync(() => precision.enabled ? fileSystemLayer : original.layer)) }
+})
+
+describe("BunHttpPlatform precision", { concurrent: false }, () => {
+  const slice = vi.fn((_start?: number, _end?: number) => new Blob([]))
+
+  beforeEach(() => {
+    precision.enabled = true
+    slice.mockClear()
+    vi.spyOn(Bun, "file").mockReturnValue({ slice } as unknown as ReturnType<typeof Bun.file>)
+  })
+  afterEach(() => {
+    precision.enabled = false
+    vi.restoreAllMocks()
+  })
+
+  testFileResponsePrecision(
+    BunHttpPlatform.layer,
+    () => assert.deepStrictEqual(slice.mock.calls, []),
+    (start, end) => assert.deepStrictEqual(slice.mock.calls, [[start, end]])
+  )
+
+  it.effect("serves a whole oversized file without slicing it", () =>
+    Effect.gen(function*() {
+      const platform = yield* HttpPlatform.HttpPlatform
+      const response = yield* platform.fileResponse("precision.bin")
+      assert.strictEqual(response.status, 200)
+      assert.strictEqual(response.body._tag, "Raw")
+      if (response.body._tag === "Raw") {
+        assert.strictEqual(response.body.body, vi.mocked(Bun.file).mock.results[0].value)
+      }
+      assert.strictEqual(vi.mocked(Bun.file).mock.calls.length, 1)
+      assert.deepStrictEqual(slice.mock.calls, [])
     }).pipe(Effect.provide(BunHttpPlatform.layer)))
 })
